Define `BOARD = UNO` on makefile instead of `BOARD = USBKEY`.
The target board for this project is Arduino UNO (+Mega2560 and Due), not USBKEY.

This simplifies and fixes some problems with the LEDs functions because both boards (UNO and USBKEY) have LEDs on PD4 and P5 but on the UNO are active low and on the USBKEY are active high. That explains why the LED function calls had to be reversed. Schematics:

Arduino UNO schematic
https://www.arduino.cc/en/uploads/Main/arduino-uno-schematic.pdf

USBKEY schematic
http://www.atmel.com/Images/doc7627.pdf (page 10)

After the call to LEDs_Init() it's not necessary to turn LEDs OFF because LEDs_Init() takes care of that as can be seen here:
https://github.com/abcminiuser/lufa/blob/master/LUFA/Drivers/Board/AVR8/UNO/LEDs.h#L89