Remove references to LEDS_LED3 and LEDS_LED4 because those LEDs do not exist on the board.
In fact, when BOARD was still difined as USBKEY, LEDS_LED4 was controlling PD7 (as can be seen on the following link) which is connected to the main AVR reset. That's tricky/risky. Was that intended?
https://github.com/abcminiuser/lufa/blob/master/LUFA/Drivers/Board/AVR8/USBKEY/LEDs.h#L81

The LED handling in EVENT_USB_Device_Connect() (and similar functions) is not too useful anyway because the LEDs are overriden on the main loop.